Moving beamintersect functions to the acquisition module because beamcirc now required knowledge of what a Beam is which would make geometry no longer selfcontained.
